+++ AzureCommunicationUI/AzureCommunicationUI/Presentation/SwiftUI/Calling/Grid/ParticipantGridLayoutView.swift
@@ -7,7 +7,8 @@ import SwiftUI
 
 struct ParticipantGridLayoutView: View {
     var cellViewModels: [ParticipantGridCellViewModel]
-    let getRemoteParticipantRendererView: (RemoteParticipantVideoViewId) -> UIView?
+    let getRemoteParticipantRendererView: (RemoteParticipantVideoViewId) -> ParticipantRendererViewInfo?
+    let rendererViewManager: RendererViewManager?
     let screenSize: ScreenSizeClassType
     let gridsMargin: CGFloat = 3
 
@@ -54,7 +55,8 @@ struct ParticipantGridLayoutView: View {
     func getRowView(cellsViewModel: [ParticipantGridCellViewModel]) -> some View {
         return ForEach(cellsViewModel) { vm in
             ParticipantGridCellView(viewModel: vm,
-                                    getRemoteParticipantRendererView: getRemoteParticipantRendererView)
+                                    getRemoteParticipantRendererView: getRemoteParticipantRendererView,
+                                    rendererViewManager: rendererViewManager)
                 .frame(maxWidth: .infinity, maxHeight: .infinity)
                 .background(Color(StyleProvider.color.surface))
                 .clipShape(RoundedRectangle(cornerRadius: 4))