+++ AzureCommunicationUI/AzureCommunicationUITests/Presentation/Calling/CallingViewModelTests.swift
@@ -157,7 +157,7 @@ class CallingViewModelTests: XCTestCase {
     func test_callingViewModel_receive_when_statusUpdated_then_participantGridViewModelUpdated() {
         let expectation = XCTestExpectation(description: "ParticipantGridViewModel is updated")
         let appState = AppState()
-        let sut = makeSUT(updateParticipantGridViewModel: { remoteParticipantsState in
+        let sut = makeSUT(updateParticipantGridViewModel: { remoteParticipantsState, _ in
             XCTAssertEqual(appState.remoteParticipantsState.lastUpdateTimeStamp, remoteParticipantsState.lastUpdateTimeStamp)
             expectation.fulfill()
         })
@@ -223,7 +223,7 @@ extension CallingViewModelTests {
                                 store: storeFactory.store)
     }
 
-    func makeSUT(updateParticipantGridViewModel: @escaping ((RemoteParticipantsState) -> Void)) -> CallingViewModel {
+    func makeSUT(updateParticipantGridViewModel: @escaping ((RemoteParticipantsState, LifeCycleState) -> Void)) -> CallingViewModel {
         let storeFactory = StoreFactoryMocking()
         let factoryMocking = CompositeViewModelFactoryMocking(logger: logger, store: storeFactory.store)
         factoryMocking.participantGridViewModel = ParticipantGridViewModelMocking(compositeViewModelFactory: factoryMocking,
@@ -242,4 +242,21 @@ extension CallingViewModelTests {
                                 logger: logger,
                                 store: storeFactory.store)
     }
+
+    func test_callingViewModel_update_when_callStatusIsConnected_appStateForeground_then_switchToBackground_shouldBecomeBackground() {
+        let sut = makeSUT()
+        let appState = AppState(callingState: CallingState(status: .connected),
+                                lifeCycleState: LifeCycleState(currentStatus: .background))
+        sut.receive(appState)
+        XCTAssertEqual(sut.appState, .background)
+    }
+
+    func test_callingViewModel_update_when_callStatusIsConnected_appStateBackground_then_switchToForeground_shouldBecomeForeground() {
+        let sut = makeSUT()
+        let appState = AppState(callingState: CallingState(status: .connected),
+                                lifeCycleState: LifeCycleState(currentStatus: .foreground))
+        sut.receive(appState)
+        XCTAssertEqual(sut.appState, .foreground)
+    }
+
 }